[Tests] Moved Float, Integer, String Validator tests to proper namespace
